feat: Adiciona funcionalidades para gerenciamento de Codaf Suplementar feat: Implementa atualização de Codaf Suplementar feat: Adiciona exclusão de Codaf Suplementar feat: Implementa geração de arquivo de remessa de conclusão feat: Adiciona upload de anexos temporários para Codaf Suplementar feat: Implementa remoção de retificações de Codaf Suplementar feat: Adiciona gerenciamento de anexos para Codaf Suplementar fix: Corrige mapeamento de DTOs para Codaf Suplementar fix: Ajusta tipos de dados em tabelas do banco de dados

+++ scripts/V114__CRIAR_TABELAS_PRINCIPAIS_CODAF_SUPLEMENTAR.sql
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS codaf_suplementar_inscricao(
 );
 
 CREATE TABLE IF NOT EXISTS codaf_suplementar_retificacao(
-	id int NOT NULL GENERATED ALWAYS AS IDENTITY(NO MINVALUE NO MAXVALUE NO CYCLE),
+	id int8 NOT NULL GENERATED ALWAYS AS IDENTITY(NO MINVALUE NO MAXVALUE NO CYCLE),
 	codaf_suplementar_id int8 NOT NULL,
 	data_retificacao date NOT NULL,
 	pagina_retificacao_dom SMALLINT NOT NULL,
